packages/app_01_forum/requirements-dev.txt
pytest-django==4.5.2
pytest-playwright==0.3.0